refactor: calendar component
