+++ workspaces/client/app/components/feature/calendar/types.ts
@@ -1,0 +1,23 @@
+export type TCalendarState = {
+	year: number;
+	month: number;
+};
+
+export type TCalendarCell =
+	| {
+			type: "day";
+			year: number;
+			month: number;
+			day: number;
+			idx: number;
+			label?: string;
+	  }
+	| {
+			type: "empty";
+			idx: number;
+	  };
+
+export type TCalendarRow = {
+	id: string;
+	cells: TCalendarCell[];
+};
